+++ pyctr/crypto/engine.py
@@ -39,7 +39,7 @@ if TYPE_CHECKING:
 
 __all__ = ['MIN_TICKET_SIZE', 'CryptoError', 'OTPLengthError', 'CorruptBootromError', 'KeyslotMissingError',
            'TicketLengthError', 'BootromNotFoundError', 'CorruptOTPError', 'Keyslot', 'CryptoEngine', 'CTRFileIO',
-           'CBCFileIO']
+           'TWLCTRFileIO', 'CBCFileIO']
 
 BOOT9_PROT_HASH = '7331f7edece3dd33f2ab4bd0b3a5d607229fd19212c10b734cedcaf78c1a7b98'