+++ pyctr/crypto/engine.py
@@ -161,6 +161,12 @@ class Keyslot(IntEnum):
     New3DSKeySector = 0x43
     """Used to decrypt the secret key sector (sector 0x96) for the New Nintendo 3DS."""
 
+    NCCHExtraKey = 0x44
+    """
+    Stores a version of another keyslot used for NCCH titles. For titles without a seed, KeyY is taken from the NCCH
+    header. For titles with a seed, KeyY is seeded. KeyX is always the same as the source keyslot.
+    """
+
 
 common_key_y = (
     # eShop