+++ pyctr/__init__.py
@@ -11,5 +11,5 @@ __copyright__ = 'Copyright (c) 2017-2021 Ian Burgwin'
 __license__ = 'MIT'
 
 VersionInfo = namedtuple('VersionInfo', 'major minor micro releaselevel serial')
-version_info = VersionInfo(major=0, minor=4, micro=6, releaselevel='beta', serial=1)
-__version__ = '0.4.7b1'
+version_info = VersionInfo(major=0, minor=4, micro=6, releaselevel='final', serial=0)
+__version__ = '0.4.7'