+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+* Use absolute paths in `CDNReader`
+
 ## v0.4.6 - March 1, 2021
 * Add pycryptodomex version requiremenet range (`>=3.9,<4`)
 * Fix using bytes file paths for `CDNReader` and `SDFilesystem`