+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 * Use absolute paths in `CDNReader`
+* Use absolute paths in `SDFilesystem`
 
 ## v0.4.6 - March 1, 2021
 * Add pycryptodomex version requiremenet range (`>=3.9,<4`)