+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 * Use absolute paths in `CDNReader`
 * Use absolute paths in `SDFilesystem`
+* Make `SubsectionIO` objects hashable (if the underlying file object is)
 
 ## v0.4.6 - March 1, 2021
 * Add pycryptodomex version requiremenet range (`>=3.9,<4`)