+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make sure two different `CTRFileIO`, `TWLCTRFileIO`, and `CBCFileIO` return different hashes, even with the same reader + key + iv/counter
 * Add `TWLCTRFileIO` to `crypto.engine.__all__`
 * Use a `frozenset` on a closed `CDNReader` object's internal open files set
+* Don't set `__del__` directly to `close` in `TypeReaderBase`, in case `close` is overridden
 
 ## v0.4.6 - March 1, 2021
 * Add pycryptodomex version requiremenet range (`>=3.9,<4`)